move metrics to proper place
