+++ eth1/goeth/goETH.go
@@ -3,47 +3,28 @@ package goeth
 import (
 	"context"
 	"fmt"
+	"math/big"
+	"strings"
+	"time"
+
 	"github.com/bloxapp/ssv/eth1"
 	"github.com/bloxapp/ssv/monitoring/metrics"
 	"github.com/bloxapp/ssv/utils/tasks"
+
 	"github.com/ethereum/go-ethereum"
 	"github.com/ethereum/go-ethereum/accounts/abi"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/ethereum/go-ethereum/ethclient"
 	"github.com/pkg/errors"
-	"github.com/prometheus/client_golang/prometheus"
-	"github.com/prometheus/client_golang/prometheus/promauto"
 	"github.com/prysmaticlabs/prysm/async/event"
 	"go.uber.org/zap"
-	"log"
-	"math/big"
-	"strings"
-	"time"
 )
 
 const (
 	healthCheckTimeout        = 10 * time.Second
 	blocksInBatch      uint64 = 100000
 )
-
-type eth1NodeStatus int32
-
-var (
-	metricsEth1NodeStatus = promauto.NewGauge(prometheus.GaugeOpts{
-		Name: "ssv:eth1:node_status",
-		Help: "Status of the connected eth1 node",
-	})
-	statusUnknown eth1NodeStatus = 0
-	statusSyncing eth1NodeStatus = 1
-	statusOK      eth1NodeStatus = 2
-)
-
-func init() {
-	if err := prometheus.Register(metricsEth1NodeStatus); err != nil {
-		log.Println("could not register prometheus collector")
-	}
-}
 
 // ClientOptions are the options for the client
 type ClientOptions struct {
@@ -140,16 +121,16 @@ func (ec *eth1Client) HealthCheck() []string {
 	defer cancel()
 	sp, err := ec.conn.SyncProgress(ctx)
 	if err != nil {
-		metricsEth1NodeStatus.Set(float64(statusUnknown))
+		reportNodeStatus(statusUnknown)
 		return []string{"could not get eth1 node sync progress"}
 	}
 	if sp != nil {
-		metricsEth1NodeStatus.Set(float64(statusSyncing))
+		reportNodeStatus(statusSyncing)
 		return []string{fmt.Sprintf("eth1 node is currently syncing: starting=%d, current=%d, highest=%d",
 			sp.StartingBlock, sp.CurrentBlock, sp.HighestBlock)}
 	}
 	// eth1 node is connected and synced
-	metricsEth1NodeStatus.Set(float64(statusOK))
+	reportNodeStatus(statusOK)
 	return []string{}
 }
