+++ eth1/goeth/metrics.go
@@ -6,18 +6,30 @@ import (
 	"log"
 )
 
+type eth1NodeStatus int32
+
 var (
 	metricSyncEventsCountSuccess = promauto.NewCounterVec(prometheus.CounterOpts{
-		Name: "ssv:eth1:sync:events:validator:count:success",
+		Name: "ssv:eth1:sync:count:success",
 		Help: "Count succeeded eth1 sync events",
 	}, []string{"etype", "self"})
 	metricSyncEventsCountFailed = promauto.NewCounterVec(prometheus.CounterOpts{
-		Name: "ssv:eth1:sync:events:validator:count:failed",
+		Name: "ssv:eth1:sync:count:failed",
 		Help: "Count failed eth1 sync events",
 	}, []string{"etype", "self"})
+	metricsEth1NodeStatus = promauto.NewGauge(prometheus.GaugeOpts{
+		Name: "ssv:eth1:node_status",
+		Help: "Status of the connected eth1 node",
+	})
+	statusUnknown eth1NodeStatus = 0
+	statusSyncing eth1NodeStatus = 1
+	statusOK      eth1NodeStatus = 2
 )
 
 func init() {
+	if err := prometheus.Register(metricsEth1NodeStatus); err != nil {
+		log.Println("could not register prometheus collector")
+	}
 	if err := prometheus.Register(metricSyncEventsCountSuccess); err != nil {
 		log.Println("could not register prometheus collector")
 	}
@@ -36,4 +48,8 @@ func reportSyncEvent(eventType string, isOperatorEvent bool, err error) {
 		return
 	}
 	metricSyncEventsCountSuccess.WithLabelValues(eventType, self).Inc()
+}
+
+func reportNodeStatus(status eth1NodeStatus) {
+	metricsEth1NodeStatus.Set(float64(status))
 }
